requirements.txt
torch
torchvision
transformers
datasets
Pillow
accelerate